backend/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "beautifulsoup4>=4.13.4",
    "chromadb>=0.6.3",
    "duckduckgo-search>=8.0.3",
    "elevenlabs>=2.3.0",
    "fastapi>=0.115.12",
    "folium>=0.19.7",
    "ftfy>=6.3.1",
    "fuzzywuzzy>=0.18.0",
    "geopandas>=1.1.0",
    "google-adk>=1.2.1",
    "google-cloud-aiplatform>=1.97.0",
    "google-genai>=1.20.0",
    "google-search-results>=2.4.2",
    "gradio>=5.33.2",
    "graphviz>=0.20.3",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "kagglehub>=0.3.12",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-docling>=1.0.0",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.25",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-mistralai>=0.2.10",
    "langchain-mongodb>=0.6.2",
    "langchain-openai>=0.3.23",
    "langchain-tavily>=0.2.2",
    "langgraph>=0.4.8",
    "langserve>=0.3.1",
    "lxml>=5.4.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.6",
    "plotly>=6.1.2",
    "pydantic>=2.11.5",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.0",
    "pypdfium2>=4.30.1",
    "python-docx>=1.1.2",
    "docx2txt>=0.8",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "regex>=2024.11.6",
    "requests>=2.32.4",
    "selenium>=4.33.0",
    "sentence-transformers>=4.1.0",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.41",
    "sse-starlette>=2.3.6",
    "streamlit>=1.45.1",
    "streamlit-agraph>=0.0.45",
    "textloader>=0.0.9",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "typing>=3.10.0.0",
    "unstructured[docx,pdf,xlsx]>=0.17.2",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
    "langchain-text-splitters>=0.3.4",
]
